fix discord bot api
ensure default users

+++ priv/repo/migrations/20240511122351_create_groups.exs
@@ -5,7 +5,7 @@ defmodule Mensaplan.Repo.Migrations.CreateGroups do
     create table(:groups) do
       add :name, :string
       add :avatar, :string
-      add :server_id, :integer
+      add :server_id, :bigint
       add :owner_id, references(:users)
 
       timestamps(type: :utc_datetime)
